+++ packages/components/components/elvis-accordion/src/react/elvia-accordion.tsx
@@ -2,6 +2,7 @@ import React, { FC, useEffect, useRef, useState } from 'react';
 import toolbox from '@elvia/elvis-toolbox';
 import styled from 'styled-components';
 import * as ElviaColors from '@elvia/elvis-colors';
+import * as ElviaTypography from '@elvia/elvis-typography';
 
 export type AccordionLabelPosition = 'left' | 'center' | 'right';
 export type AccordionSize = 'small' | 'medium' | 'large';
@@ -20,6 +21,10 @@ const colors = {
   green: ElviaColors.default['primary-colors']['green'].color,
   white: ElviaColors.default['primary-colors']['white'].color,
   black: ElviaColors.default['primary-colors']['black'].color,
+};
+
+export const typography = {
+  textLabel: ElviaTypography.default['text-label'],
 };
 
 const AccordionArea = styled.div`
@@ -82,8 +87,8 @@ const AccordionButton = styled.button<AccordionButton>`
   background: transparent;
   display: flex;
   padding: 0;
-  font-family: 'Red Hat Display', Verdana, sans-serif;
-  font-weight: 500;
+  font-family: ${typography.textLabel.fontFamily};
+  font-weight: ${typography.textLabel.fontWeight};
   font-size: ${(props: { size: AccordionSize }) => decideButtonFontSize(props.size)};
   line-height: ${(props: { size: AccordionSize }) => (props.size === 'small' ? '16px' : '24px')};
   text-align: left;