+++ packages/components/components/elvis-dropdown/CHANGELOG.md
@@ -1,10 +1,11 @@
 # Elvia Dropdown
 
-## 1.0.9 (19.11.21)
+## 1.0.9 (23.11.21)
 
 ### Bug fix
 
 - Now using elvis-colors pack for colors.
+- Now using elvis-typography pack for typography.
 
 ## 1.0.8 (28.09.21)